+++ Server/Lib/Net/ConnectionMUDP.cpp
@@ -1019,7 +1019,7 @@ namespace Net {
 		pIOBuffer->SetupRecvUDP(GetCID());
 
 		hrErr = NetSystem::RecvFrom(GetSocket(), pIOBuffer);
-		switch ((int32_t)hrErr)
+		switch ((int64_t)hrErr)
 		{
 		case ResultCode::SUCCESS_FALSE:
 			hr = ResultCode::E_NET_TRY_AGAIN;
@@ -1136,7 +1136,7 @@ namespace Net {
 		{
 			pOver = new IOBUFFER_READ;
 			hrErr = Recv(pOver);
-			switch ((int32_t)hrErr)
+			switch ((uint32_t)hrErr)
 			{
 			case ResultCode::SUCCESS:
 			case ResultCode::E_NET_IO_PENDING:
